+++ Firmware/FFBoard/UserExtensions/Inc/TMC4671.h
@@ -90,6 +90,7 @@ struct TMC4671HardwareTypeConf{
 	uint16_t brakeLimHigh = 50900;
 	float vmScaler = (2.5 / 0x7fff) * ((1.5+71.5)/1.5);
 	float vSenseMult = VOLTAGE_MULT_DEFAULT;
+	float clockfreq = 25e6;
 	// Todo restrict allowed motor and encoder types
 };
 
@@ -139,7 +140,7 @@ union StatusFlags {
 struct TMC4671MainConfig{
 	TMC4671HardwareTypeConf hwconf;
 	TMC4671MotConf motconf;
-	uint16_t pwmcnt 		= 3999;
+	uint16_t pwmcnt 		= 3999; // PWM resolution is 12 bit internally
 	uint8_t bbmL			= 10;
 	uint8_t bbmH			= 10;
 	uint16_t mdecA 			= 660; // 334 default. 331 recommended by datasheet,662 double. 660 lowest noise
@@ -201,6 +202,7 @@ struct TMC4671FlashAddrs{
 	uint16_t ADC_i1_ofs = ADR_TMC1_ADC_I1_OFS;
 	uint16_t encOffset = ADR_TMC1_ENC_OFFSET;
 	uint16_t phieOffset = ADR_TMC1_PHIE_OFS;
+	uint16_t torqueFilter = ADR_TMC1_TRQ_FILT;
 };
 
 struct TMC4671ABNConf{
@@ -258,26 +260,39 @@ struct TMC4671HALLConf{
 	uint16_t dPhiMax = 10922;
 };
 
+enum class TMCbiquadpreset : uint8_t {none=0,lowpass=1,notch=2,peak=3};
 // Use TMCL IDE with TMC4671 devkit to generate values
-// TODO rewrite as class to allow conversion from biquad
-struct TMC4671Biquad{
-
+struct TMC4671Biquad_t{
 	int32_t a1 = 0;
 	int32_t a2 = 0;
 	int32_t b0 = 0;
 	int32_t b1 = 0;
 	int32_t b2 = 0;
 	bool enable = false;
+};
+struct TMC4671Biquad_conf{
+	TMCbiquadpreset mode;
+	biquad_constant_t params = {1000,50}; // Q = 1/100 for lowpass and 1/10 for notch and peak mode
+	float gain = 10.0; // Gain for peak mode
+};
 
-	void from_bq(Biquad& bq,bool enable){
-		// Note: trinamic swapped the naming of b and a from the regular convention in the datasheet
-		a1 = (int32_t)(bq.b1 * (1 << 29));
-		a2 = (int32_t)(bq.b2 * (1 << 29));
-		b0 = (int32_t)(bq.a0 * (1 << 29));
-		b1 = (int32_t)(bq.a1 * (1 << 29));
-		b2 = (int32_t)(bq.a2 * (1 << 29));
-		this->enable = enable;
-	};
+class TMC4671Biquad{
+public:
+	TMC4671Biquad(bool enable = false){
+		params.enable = enable;
+	}
+	TMC4671Biquad(const TMC4671Biquad_t bq) : params(bq){}
+	TMC4671Biquad(const Biquad& bq,bool enable = true){
+		// Note: trinamic swapped the naming of b and a from the regular convention in the datasheet and a and b are possibly inverse to b in our filter class
+		this->params.a1 = -(int32_t)(bq.b1 * (float)(1 << 29));
+		this->params.a2 = -(int32_t)(bq.b2 * (float)(1 << 29));
+		this->params.b0 = (int32_t)(bq.a0 * (float)(1 << 29));
+		this->params.b1 = (int32_t)(bq.a1 * (float)(1 << 29));
+		this->params.b2 = (int32_t)(bq.a2 * (float)(1 << 29));
+		this->params.enable = bq.getFc() > 0 ? enable : false;
+	}
+
+	TMC4671Biquad_t params;
 };
 
 
@@ -296,7 +311,7 @@ class TMC4671 :
 		torqueP,torqueI,fluxP,fluxI,velocityP,velocityI,posP,posI,
 		tmctype,pidPrec,phiesrc,fluxoffset,seqpi,tmcIscale,encdir,temp,reg,
 		svpwm,fullCalibration,calibrated,abnindexenabled,findIndex,getState,encpol,combineEncoder,invertForce,vmTmc,
-		extphie
+		extphie,torqueFilter_mode,torqueFilter_f,torqueFilter_q
 	};
 
 public:
@@ -365,11 +380,11 @@ public:
 	void setFFMode(FFMode mode);
 	void setSequentialPI(bool sequential);
 
-	void setBiquadFlux(TMC4671Biquad bq);
-	void setBiquadTorque(TMC4671Biquad bq);
-	void setBiquadPos(TMC4671Biquad bq);
-	void setBiquadVel(TMC4671Biquad bq);
-	
+	void setBiquadFlux(const TMC4671Biquad &filter);
+	void setBiquadTorque(const TMC4671Biquad &filter);
+	void setBiquadPos(const TMC4671Biquad &filter);
+	void setBiquadVel(const TMC4671Biquad &filter);;
+	void setTorqueFilter(TMC4671Biquad_conf& conf);
 
 	bool pingDriver();
 	std::pair<uint32_t,std::string> getTmcType();
@@ -379,8 +394,7 @@ public:
 	bool externalEncoderAllowed();
 	void setExternalEncoderAllowed(bool allow);
 
-	bool isCalibrated();
-
+	float getPwmFreq();
 
 
 #ifdef TIM_TMC
@@ -592,6 +606,10 @@ private:
 	uint32_t initTime = 0;
 	bool manualEncAlign = false;
 	bool spiActive = false; // Flag for tx interrupt that the transfer was started by this instance
+
+	TMCbiquadpreset torqueFilterMode = TMCbiquadpreset::none;
+
+	TMC4671Biquad_conf torqueFilterConf;
 
 	// External encoder timer fires interrupts to trigger a new commutation position update
 #ifdef TIM_TMC