+++ Firmware/FFBoard/UserExtensions/Inc/eeprom_addresses.h
@@ -12,11 +12,11 @@
 		
 #include "main.h"		
 // Change this to the amount of currently registered variables		
-#define NB_OF_VAR 141
+#define NB_OF_VAR 144
 extern const uint16_t VirtAddVarTab[NB_OF_VAR];		
 		
 // Amount of variables in exportable list		
-#define NB_EXPORTABLE_ADR 126
+#define NB_EXPORTABLE_ADR 129
 extern const uint16_t exportableFlashAddresses[NB_EXPORTABLE_ADR];		
 		
 		
@@ -109,6 +109,7 @@ uint16_t EE_ReadVariable(uint16_t VirtAddress, uint16_t* Data) will return 1 if 
 #define ADR_TMC1_FLUX_P 0x329 		
 #define ADR_TMC1_FLUX_I 0x32A 		
 #define ADR_TMC1_PHIE_OFS 0x32B 		
+#define ADR_TMC1_TRQ_FILT 0x32C
 // AXIS2		
 #define ADR_AXIS2_CONFIG 0x341 // 0-2 ENC, 3-5 DRV		
 #define ADR_AXIS2_POWER 0x342 		
@@ -132,6 +133,7 @@ uint16_t EE_ReadVariable(uint16_t VirtAddress, uint16_t* Data) will return 1 if 
 #define ADR_TMC2_FLUX_P 0x369 		
 #define ADR_TMC2_FLUX_I 0x36A 		
 #define ADR_TMC2_PHIE_OFS 0x36B 		
+#define ADR_TMC2_TRQ_FILT 0x36C
 // AXIS3		
 #define ADR_AXIS3_CONFIG 0x381 // 0-2 ENC, 3-5 DRV		
 #define ADR_AXIS3_POWER 0x382 		
@@ -155,6 +157,7 @@ uint16_t EE_ReadVariable(uint16_t VirtAddress, uint16_t* Data) will return 1 if 
 #define ADR_TMC3_FLUX_P 0x3A9 		
 #define ADR_TMC3_FLUX_I 0x3AA 		
 #define ADR_TMC3_PHIE_OFS 0x3AB 		
+#define ADR_TMC3_TRQ_FILT 0x3AC
 // Odrive		
 #define ADR_ODRIVE_CANID 0x3D0 //0-6 ID M0, 7-12 ID M1, 13-15 can speed		
 #define ADR_ODRIVE_SETTING1_M0 0x3D1