+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 ### Changes since 1.9.7:
 - Fixed an issue with 2 axis FFB effects on second axis
+- Added TMC4671 biquad filter option
+  - Lowpass, notch and peak modes (fixed Q factor, saved frequency)